seq772 - Comments added indirectly propagate to requests
